requirements.txt
django-model-utils>=2.0
# Additional requirements go here
beautifulsoup4>=4.6.0
requests>=2.18.1
